daily-wisdom-api/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
greenlet==3.0.3